chore(builder): move getPostcssConfig to shared
