+++ packages/builder/builder-rspack-provider/src/plugins/css.ts
@@ -1,4 +1,3 @@
-import assert from 'assert';
 import {
   getBrowserslistWithDefault,
   isUseCssSourceMap,
@@ -7,12 +6,11 @@ import {
   type BuilderContext,
   BundlerChain,
   ModifyBundlerChainUtils,
-  getCssSupport,
   setConfig,
   logger,
-  PostCSSPlugin,
   CssModules,
   getCssModulesAutoRule,
+  getPostcssConfig,
 } from '@modern-js/builder-shared';
 import type {
   BuilderPlugin,
@@ -21,23 +19,6 @@ import type {
   RuleSetRule,
 } from '../types';
 import { isUseCssExtract, getCompiledPath } from '../shared';
-
-type CssNanoOptions = {
-  configFile?: string | undefined;
-  preset?: [string, object] | string | undefined;
-};
-
-export const getCssnanoDefaultOptions = (): CssNanoOptions => ({
-  preset: [
-    'default',
-    {
-      // merge longhand will break safe-area-inset-top, so disable it
-      // https://github.com/cssnano/cssnano/issues/803
-      // https://github.com/cssnano/cssnano/issues/967
-      mergeLonghand: false,
-    },
-  ],
-});
 
 export async function applyBaseCSSRule(
   rule: ReturnType<BundlerChain['module']['rule']>,
@@ -50,71 +31,13 @@ export async function applyBaseCSSRule(
   const enableSourceMap = isUseCssSourceMap(config);
   // const enableCSSModuleTS = Boolean(config.output.enableCssModuleTSDeclaration);
 
-  const { applyOptionsChain } = await import('@modern-js/utils');
   const browserslist = await getBrowserslistWithDefault(
     context.rootPath,
     config,
     target,
   );
 
-  // todo: move to modern-js/shared
-  const getPostcssConfig = () => {
-    const extraPlugins: PostCSSPlugin[] = [];
-    const utils = {
-      addPlugins(plugins: PostCSSPlugin | PostCSSPlugin[]) {
-        if (Array.isArray(plugins)) {
-          extraPlugins.push(...plugins);
-        } else {
-          extraPlugins.push(plugins);
-        }
-      },
-    };
-    const enableCssMinify = !enableExtractCSS && isProd;
-
-    const cssSupport = getCssSupport(browserslist);
-
-    const mergedConfig = applyOptionsChain(
-      {
-        postcssOptions: {
-          plugins: [
-            require(getCompiledPath('postcss-flexbugs-fixes')),
-            !cssSupport.customProperties &&
-              require(getCompiledPath('postcss-custom-properties')),
-            !cssSupport.initial && require(getCompiledPath('postcss-initial')),
-            !cssSupport.pageBreak &&
-              require(getCompiledPath('postcss-page-break')),
-            !cssSupport.fontVariant &&
-              require(getCompiledPath('postcss-font-variant')),
-            !cssSupport.mediaMinmax &&
-              require(getCompiledPath('postcss-media-minmax')),
-            require(getCompiledPath('postcss-nesting')),
-            require(getCompiledPath('autoprefixer'))(
-              applyOptionsChain(
-                {
-                  flexbox: 'no-2009',
-                  overrideBrowserslist: browserslist,
-                },
-                config.tools.autoprefixer,
-              ),
-            ),
-            enableCssMinify
-              ? require('cssnano')(getCssnanoDefaultOptions())
-              : false,
-          ].filter(Boolean),
-        },
-        sourceMap: enableSourceMap,
-      },
-      // postcss-loader will modify config
-      config.tools.postcss || {},
-      utils,
-    );
-    if (extraPlugins.length) {
-      assert('postcssOptions' in mergedConfig);
-      assert('plugins' in mergedConfig.postcssOptions!);
-      mergedConfig.postcssOptions.plugins!.push(...extraPlugins);
-    }
-    return mergedConfig;
-  };
+  const enableCssMinify = !enableExtractCSS && isProd;
 
   /**
    * TODO: support style-loader & ignore css (need Rspack support inline css first)
@@ -141,7 +64,12 @@ export async function applyBaseCSSRule(
   // }
 
   if (!isServer && !isWebWorker) {
-    const postcssLoaderOptions = getPostcssConfig();
+    const postcssLoaderOptions = await getPostcssConfig({
+      enableSourceMap,
+      browserslist,
+      config,
+      enableCssMinify,
+    });
 
     rule
       .use(CHAIN_ID.USE.POSTCSS)
