+++ packages/builder/builder-webpack-provider/package.json
@@ -98,7 +98,6 @@
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.9",
     "caniuse-lite": "^1.0.30001489",
     "css-minimizer-webpack-plugin": "5.0.0",
-    "cssnano": "6.0.0",
     "html-webpack-plugin": "5.5.0",
     "line-diff": "^2.0.1",
     "mini-css-extract-plugin": "2.7.0",
