+++ packages/builder/builder-webpack-provider/src/plugins/css.ts
@@ -1,8 +1,7 @@
 import path from 'path';
-import assert from 'assert';
 import {
   CSS_REGEX,
-  getCssSupport,
+  getPostcssConfig,
   ModifyChainUtils,
   isUseCssSourceMap,
   getCssModulesAutoRule,
@@ -19,8 +18,6 @@ import type {
   NormalizedConfig,
   StyleLoaderOptions,
 } from '../types';
-import type { AcceptedPlugin, ProcessOptions } from 'postcss';
-import { getCssnanoDefaultOptions } from './minimize';
 
 export const isUseCssExtract = (
   config: NormalizedConfig,
@@ -81,73 +78,10 @@ export async function applyBaseCSSRule(
     target,
   );
 
-  // todo: move to modern-js/shared
-  const getPostcssConfig = () => {
-    const extraPlugins: AcceptedPlugin[] = [];
-
-    const utils = {
-      addPlugins(plugins: AcceptedPlugin | AcceptedPlugin[]) {
-        if (Array.isArray(plugins)) {
-          extraPlugins.push(...plugins);
-        } else {
-          extraPlugins.push(plugins);
-        }
-      },
-    };
-
-    const enableCssMinify = !enableExtractCSS && isProd;
-
-    const cssSupport = getCssSupport(browserslist);
-
-    const mergedConfig = applyOptionsChain(
-      {
-        postcssOptions: {
-          plugins: [
-            require(getCompiledPath('postcss-flexbugs-fixes')),
-            !cssSupport.customProperties &&
-              require(getCompiledPath('postcss-custom-properties')),
-            !cssSupport.initial && require(getCompiledPath('postcss-initial')),
-            !cssSupport.pageBreak &&
-              require(getCompiledPath('postcss-page-break')),
-            !cssSupport.fontVariant &&
-              require(getCompiledPath('postcss-font-variant')),
-            !cssSupport.mediaMinmax &&
-              require(getCompiledPath('postcss-media-minmax')),
-            require(getCompiledPath('postcss-nesting')),
-            require(getCompiledPath('autoprefixer'))(
-              applyOptionsChain(
-                {
-                  flexbox: 'no-2009',
-                  overrideBrowserslist: browserslist,
-                },
-                config.tools.autoprefixer,
-              ),
-            ),
-            enableCssMinify
-              ? require('cssnano')(getCssnanoDefaultOptions())
-              : false,
-          ].filter(Boolean),
-        },
-        sourceMap: enableSourceMap,
-      },
-      config.tools.postcss || {},
-      utils,
-    );
-    if (extraPlugins.length) {
-      assert('postcssOptions' in mergedConfig);
-      assert('plugins' in mergedConfig.postcssOptions!);
-      mergedConfig.postcssOptions.plugins!.push(...extraPlugins);
-    }
-
-    return mergedConfig as ProcessOptions & {
-      postcssOptions: {
-        plugins?: AcceptedPlugin[];
-      };
-    };
-  };
-
   // 1. Check user config
   const enableExtractCSS = isUseCssExtract(config, target);
+  const enableCssMinify = !enableExtractCSS && isProd;
+
   const enableSourceMap = isUseCssSourceMap(config);
   const enableCSSModuleTS = Boolean(config.output.enableCssModuleTSDeclaration);
   // 2. Prepare loader options
@@ -238,7 +172,13 @@ export async function applyBaseCSSRule(
     .end();
 
   if (!isServer && !isWebWorker) {
-    const postcssLoaderOptions = getPostcssConfig();
+    const postcssLoaderOptions = await getPostcssConfig({
+      enableSourceMap,
+      browserslist,
+      config,
+      enableCssMinify,
+    });
+
     rule
       .use(CHAIN_ID.USE.POSTCSS)
       .loader(getCompiledPath('postcss-loader'))
